+++ examples/Steps.ps1
@@ -1,7 +1,7 @@
 ﻿<#
 .Synopsis
-	Runs Steps.stateless.ps1 by Invoke-StateMachine
+	Runs Steps.stateless.ps1 by Invoke-StateMachine.
 #>
 
 $machine = & $PSScriptRoot\Steps.stateless.ps1
-Invoke-StateMachine $machine -Show
+Invoke-StateMachine $machine -AddShow