+++ examples/csharp/BugTracker.ps1
@@ -3,4 +3,4 @@ Add-Type -Path $PSScriptRoot\bin\Debug\net8.0\Extra.dll
 
 $test = [Extra.BugTracker]::new('The bug', {Write-Host $args})
 
-Invoke-StateMachine $test.Machine -Show
+Invoke-StateMachine $test.Machine -AddShow