+++ tests/EditFlags.fas.ps1
@@ -1,0 +1,34 @@
+﻿#! Covers unwrapping PSObject machine created by New-Object.
+
+Import-Module $PSScriptRoot/About.psm1
+
+run {
+	$r = & $PSScriptRoot\..\examples\EditFlags.ps1
+	Assert-Far "$r" -eq 'Alt, Control'
+}
+
+job {
+	Assert-Far $Far.Dialog[0].Text -eq None
+}
+
+keys Enter
+job {
+	Assert-Far $Far.Dialog[0].Text -eq Alt
+}
+
+keys Enter
+job {
+	Assert-Far $Far.Dialog[0].Text -eq 'Alt, Shift'
+}
+
+keys Enter
+job {
+	Assert-Far $Far.Dialog[0].Text -eq 'Alt, Shift, Control'
+}
+
+keys 2
+job {
+	Assert-Far $Far.Dialog[0].Text -eq 'Alt, Control'
+}
+
+keys 0